# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/orders": {
      "post": {
        "operationId": "OrdersController_create",
        "parameters": [
          {
            "name": "idempotency-key",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Évite une double commande en cas de double clic",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Créer une commande PENDING et une session de paiement",
        "tags": [
          "Commandes"
        ]
      },
      "get": {
        "operationId": "OrdersController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Commandes"
        ]
      }
    },
    "/v1/orders/{id}": {
      "get": {
        "operationId": "OrdersController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Statut d'une commande (interrogé par l'app après le retour de paiement)",
        "tags": [
          "Commandes"
        ]
      }
    },
...
    "/v1/admin/orders/{id}": {
      "get": {
        "operationId": "AdminController_getOrder",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin"
        ]
      }
    },
...
  },
  "info": {
    "title": "API Folio",
    "description": "Vente et lecture d'e-books — application mobile et administration. Toutes les routes sont préfixées par /v1.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateOrderDto": {
        "type": "object",
        "properties": {
          "ebookIds": {
            "description": "Par défaut : le contenu du panier",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "platform": {
            "type": "string",
            "enum": [
              "ios",
              "android",
              "web"
            ]
          },
          "returnUrl": {
            "type": "string",
            "description": "Adresse de retour de l'app après paiement (préfixes autorisés : CHECKOUT_RETURN_PREFIXES)"
          }
        }
      },
...
    }
  }
}